fix first couple tests rename files
